Refactor the analytics module
